+++ src/Optimizely/Event/Builder/EventBuilder.php
@@ -38,7 +38,7 @@ class EventBuilder
     /**
      * @const string Version of the Optimizely PHP SDK.
      */
-    const SDK_VERSION = '3.4.0';
+    const SDK_VERSION = '3.6.1';
 
     /**
      * @var string URL to send event to.
@@ -143,14 +143,22 @@ class EventBuilder
      *
      * @return array Hash representing parameters particular to impression event.
      */
-    private function getImpressionParams(Experiment $experiment, $variationId)
+    private function getImpressionParams(Experiment $experiment, $variation, $flagKey, $ruleKey, $ruleType, $enabled)
     {
+        $variationKey = $variation->getKey() ? $variation->getKey() : '';
         $impressionParams = [
             DECISIONS => [
                 [
                     CAMPAIGN_ID => $experiment->getLayerId(),
                     EXPERIMENT_ID => $experiment->getId(),
-                    VARIATION_ID => $variationId
+                    VARIATION_ID => $variation->getId(),
+                    METADATA => [
+                        FLAG_KEY => $flagKey,
+                        RULE_KEY => $ruleKey,
+                        RULE_TYPE => $ruleType,
+                        VARIATION_KEY => $variationKey,
+                        ENABLED => $enabled
+                    ],
                 ]
             ],
 
@@ -221,13 +229,13 @@ class EventBuilder
      *
      * @return LogEvent Event object to be sent to dispatcher.
      */
-    public function createImpressionEvent($config, $experimentKey, $variationKey, $userId, $attributes)
+    public function createImpressionEvent($config, $experimentKey, $variationKey, $flagKey, $ruleKey, $ruleType, $enabled, $userId, $attributes)
     {
         $eventParams = $this->getCommonParams($config, $userId, $attributes);
 
         $experiment = $config->getExperimentFromKey($experimentKey);
         $variation = $config->getVariationFromKey($experimentKey, $variationKey);
-        $impressionParams = $this->getImpressionParams($experiment, $variation->getId());
+        $impressionParams = $this->getImpressionParams($experiment, $variation, $flagKey, $ruleKey, $ruleType, $enabled);
 
         $eventParams[VISITORS][0][SNAPSHOTS][] = $impressionParams;