+++ tests/EventTests/EventBuilderTest.php
@@ -68,7 +68,7 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
                 ]],
                 'revision' => '15',
                 'client_name' => 'php-sdk',
-                'client_version' => '3.4.0',
+                'client_version' => '3.6.1',
                 'anonymize_ip'=> false,
                 'enrich_decisions' => true,
             ];
@@ -76,7 +76,14 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
         $decisions = array('decisions' => [[
                               'campaign_id'=> '7719770039',
                               'experiment_id'=> '7716830082',
-                              'variation_id'=> '7721010009'
+                              'variation_id'=> '7721010009',
+                              'metadata'=> [
+                                  'flag_key' => 'test_experiment',
+                                  'rule_key' => 'test_experiment',
+                                  'rule_type' => 'experiment',
+                                  'variation_key'=> 'variation',
+                                  'enabled' => true
+                              ]
                             ]]
                           );
         $this->expectedImpressionEventParams = $this->expectedEventParams;
@@ -137,6 +144,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             null
         );
@@ -193,6 +204,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -230,6 +245,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -268,6 +287,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -296,6 +319,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -333,6 +360,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -377,6 +408,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -424,6 +459,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $configMock,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -467,6 +506,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $configMock,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );
@@ -793,6 +836,10 @@ class EventBuilderTest extends \PHPUnit_Framework_TestCase
             $this->config,
             'test_experiment',
             'variation',
+            'test_experiment',
+            'test_experiment',
+            'experiment',
+            true,
             $this->testUserId,
             $userAttributes
         );